requirements.txt
python-dotenv==1.0.0
gradio==4.7.1
openai==1.3.6
langchain==0.0.343
google-cloud-texttospeech==2.14.2
PyGithub==2.1.1